Community organizing/open source way books
These are all on a list at http://www.theopensourceway.org/wiki/Data_and_references and I added that link to the "inspired by these sources" section.

I hesitated to suggest The Open Source Way itself as it is an ongoing, always-a-work-in-progress, and undergoing active editing to improve. It may not be a good resource for the target audience of this list. If you take a look and want to include it, you can use this change under the Red Hat authors section

|  The Open Source Way(http://www.theopensourceway.org) | Karsten Wade and others |

... or you can let me know to add it myself and I'll make the pull request.